Added exp grid to classes where custom exp tnl's can be set :)
Former-commit-id: 8a6433b959db65707e226c6b8e76fc4bf7bd4114

+++ Intersect.Editor/Forms/Editors/Events/Event Commands/EventCommand_PlayAnimation.cs
@@ -176,8 +176,8 @@ namespace Intersect.Editor.Forms.Editors.Events.Event_Commands
                 case 0: //Tile Spawn
                     mMyCommand.EntityId = Guid.Empty;
                     mMyCommand.MapId = MapList.OrderedMaps[cmbMap.SelectedIndex].MapId;
-                    mMyCommand.X = (int) nudWarpX.Value;
-                    mMyCommand.Y = (int) nudWarpY.Value;
+                    mMyCommand.X = (sbyte) nudWarpX.Value;
+                    mMyCommand.Y = (sbyte) nudWarpY.Value;
                     mMyCommand.Dir = (byte)cmbDirection.SelectedIndex;
                     break;
                 case 1: //On/Around Entity Spawn
@@ -190,8 +190,8 @@ namespace Intersect.Editor.Forms.Editors.Events.Event_Commands
                     {
                         mMyCommand.EntityId = mCurrentMap.LocalEvents.Keys.ToList()[cmbEntities.SelectedIndex - 1];
                     }
-                    mMyCommand.X = mSpawnX;
-                    mMyCommand.Y = mSpawnY;
+                    mMyCommand.X = (sbyte)mSpawnX;
+                    mMyCommand.Y = (sbyte)mSpawnY;
                     if (chkRelativeLocation.Checked && chkRotateDirection.Checked)
                     {
                         mMyCommand.Dir = 3;
